backend/ai/requirements.txt
fastapi~=0.111.1
uvicorn~=0.30.4
openai~=1.41.0
pydantic
python-dotenv~=1.0.1
rembg~=2.0.57
pillow~=10.4.0
pydantic_core==2.20.1
colorlog==6.8.2
torch==2.4.0
transformers==4.43.3
pika
sse_starlette
matplotlib
torchvision
requests-toolbelt